Just implemented Nuke to handle Images cache and to avoid unecesary reloading.
